feat: prefetch 방식 적용을 위해 main 페이지 서버 컴포넌트에 prefetch 로직 추가
